fix: vendor build step clashed with vendor/bundle dir
point ruby's bundler to install deps in .bundle dir instead of the
vendor/bundle dir

the build system uses vendor/ to install CDN assets, so we dont want
those to collide.

specifically, previous commit had told jekyll to ignore the vendor dir
after they had pointed the bundler to install deps there. but in reality
we need that vendor dir to be part of the build for now

also ignored package-lock.json from jekyll, as it was ending up being
shipped to prod